+++ test/foo.test.js
@@ -15,10 +15,10 @@ describe('Send Buffer Data', function(){
         this.timeout(100 * 1000);
         // handshake message
         client.connect(PORT, LOCAL_HOST, function(){
-            data = Buffer.from([0x01, 02, 03, 04, 0x2b, 0xa1], 'hex')
-            client.on('data', (buf) => {
-                console.info('Receiverd:', buf)
-            })
+            data = Buffer.from([0x01, 02, 00, 00, 0x2b, 0xa1], 'hex')
+            // client.on('data', (buf) => {
+            //     console.info('Receiverd:', buf)
+            // })
             client.write(data, (e) => {
                 if(e){
                     console.info(e)
@@ -56,13 +56,19 @@ describe('Send Buffer Data', function(){
 
     it('sendData', function(done){
         var func = new YF.Func('socket.send');
-        func.invoke({id: '01023', message: '12121212'})
+        func.invoke({id: '0102', message: '010200010506', callback: '0001'})
           .then(function(d){
-            console.info(d)
+            console.info('The Return Data Of SendData:', d)
             done();
           }).catch(function(err){
             done(err);
           });
+          setTimeout(() => {
+            client.write(Buffer.from([0x01, 02, 00, 01, 0x2b, 0xa1], 'hex'), (e) => {
+                console.info('write ok')
+            })
+          }, 0.5 * 1000)
+          
     });
 
     after(function(done) {
@@ -70,8 +76,11 @@ describe('Send Buffer Data', function(){
         // runs after each test in this block
         console.log('Client: >>>> destroy');
         try{
-            client.destroy();
-            done();
+            setTimeout( () => {
+                client.destroy();
+                done();
+            }, 10 * 1000)
+            
         }catch(e){
             done(e);
         }